Circuit 3 code
this is in complete as I haven't been able
to get it go go through all 8 colors.

current problem is red is high.
blue has frequency of 3.5Hz
green has frequency of 7.2Hz

if i loop through the values it seems to work so i don't
think it's a short in the hardware but an incorrect assumption
on my part in the code.